Extend claude binary resolution to all spawn sites
- Extract resolveClaudeBin + pathWithNativeBin into shared src/session/claude-bin.ts
- compactor.ts: `claude --print` now resolves the binary and hardens PATH
- process.ts: guard app-agents to bare `claude` (fix container-path-leak regression), add broken-shim clarifying comment
- Tests: new claude-bin.test.ts + session-process and compactor additions
